chore(bukkit/nms): implement ItemDisplayNPC
