--- stickynote-bukkit/stickynote-bukkit-nms/src/main/kotlin/org/sayandev/stickynote/bukkit/nms/npc/entity/ItemDisplayNPC.kt
@@ -1,17 +1,0 @@
-package org.sayandev.stickynote.bukkit.nms.npc.entity
-
-import org.bukkit.Location
-import org.bukkit.inventory.ItemStack
-import org.sayandev.stickynote.bukkit.nms.npc.NPCType
-
-class ItemDisplayNPC(
-    location: Location,
-    item: ItemStack
-): DisplayNPC(
-    location,
-    NPCType.ITEM_DISPLAY
-) {
-
-
-
-}
